Code Integration of Anand and Kuldeep commited by Anand

+++ src/main/java/com/fin/analyzer/repository/AccountRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<AccountDetails, Integer> {
+public interface AccountRepository extends JpaRepository<AccountDetails, Long> {
 }
